Jieun code merge

+++ src/main/java/WeMatch/wematch/domain/group/service/TeamService.java
@@ -7,6 +7,8 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 
+import java.time.DayOfWeek;
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 
@@ -28,9 +30,19 @@ public class TeamService {
         Iterator<TeamEventsResponseDto> iterator = events.iterator();
         while (iterator.hasNext()) {
             TeamEventsResponseDto event = iterator.next();
-            if (event.getEventStartAt().toLocalDate().get(WeekFields.ISO.weekOfYear()) != weekNumber
-                    || event.getEventStartAt().getYear() != year) {
+            // 연도로 거름
+            if (event.getStart().getYear() != year) { //연도
                 iterator.remove();
+            }
+            // weekNumber로 거름
+            if(event.getStart().toLocalDate().getDayOfWeek()== DayOfWeek.SUNDAY) {
+                if(event.getStart().toLocalDate().get(WeekFields.ISO.weekOfYear()) != (weekNumber-1)) {
+                    iterator.remove();
+                }
+            } else {
+                if(event.getStart().toLocalDate().get(WeekFields.ISO.weekOfYear()) != weekNumber) {
+                    iterator.remove();
+                }
             }
         }
         return events;
@@ -64,30 +76,53 @@ public class TeamService {
         teamRepository.insertFixedTime(members, teamName, insertFixedTimeRequestDto);
     }
 
-    public CandidateResponseDto getCandidate(Long candidateId) {
+    public GetCandidateFixDto getCandidate(Long candidateId) {
         InsertFixedTimeRequestDto candidate = teamRepository.getCandidate(candidateId);
-        return CandidateResponseDto.builder()
+        return GetCandidateFixDto.builder()
                 .startDate(candidate.getStartAt().toLocalDate())
                 .startAt(candidate.getStartAt().toLocalTime())
                 .endAt(candidate.getEndAt().toLocalTime())
                 .build();
     }
 
-    public List<TeamEventsResponseDto> getCandidates(Long groupId) {
+    public List<TeamEventsResponseDto> getCandidates(Long groupId,int weekNumber, int year) {
+        List<CandidateResponseDto> returnDto = new ArrayList<>();
         List<TeamEventsResponseDto> results = teamRepository.getEvent(groupId);
+        // weekNumber, year로 제한
+        results = getResult(results,weekNumber,year);
+        // no results -> return
+        if(results.isEmpty()) {return results;}
+
         List<TeamEventsResponseDto> results2 = teamRepository.getEvent(groupId);
+        results2 = getResult(results2,weekNumber,year);
         //startAt으로 정렬
-        Collections.sort(results2, Comparator.comparing(TeamEventsResponseDto::getEventStartAt));
+        Collections.sort(results2, Comparator.comparing(TeamEventsResponseDto::getStart));
 
         //delete
         teamRepository.deleteCandidates(groupId);
         //getCandidates
         results = getBlankTime(results);
         results = setFirstLast(results, results2, groupId);
+        //수면시간 제거
         results = deleteSleepTime(results, groupId);
+        // 재정렬
+        Collections.sort(results, Comparator.comparing(TeamEventsResponseDto::getStart));
+        // 분단위 시간 제거
+        results = deleteMinuteTime(results,teamRepository.getMinute(groupId).getMinute());
         //insertCandidates
-        teamRepository.insertCandidates(groupId,results);
+        List<Long> candidateIds = teamRepository.insertCandidates(groupId,results);
+        addCandidateIds(results, candidateIds);
         return results;
+    }
+
+
+    private void addCandidateIds(List<TeamEventsResponseDto> results, List<Long> candidateIds) {
+        Iterator<TeamEventsResponseDto> iterator = results.iterator();
+        int i=0;
+        while (iterator.hasNext()) {
+            TeamEventsResponseDto item = iterator.next();
+            item.setTitle(candidateIds.get(i++));
+        }
     }
 
     public List<TeamEventsResponseDto> getBlankTime(List<TeamEventsResponseDto> events) {
@@ -98,26 +133,26 @@ public class TeamService {
             // 모든 시간대가 없는 경우
             missingTimeIntervals.add(
                     TeamEventsResponseDto.builder()
-                            .eventStartAt(LocalDateTime.MIN)
-                            .eventEndAt(LocalDateTime.MAX)
+                            .start(LocalDateTime.MIN)
+                            .end(LocalDateTime.MAX)
                             .build());
             return missingTimeIntervals;
         }
 
-        events.sort((event1, event2) -> event1.getEventStartAt().compareTo(event2.getEventStartAt()));
+        events.sort((event1, event2) -> event1.getStart().compareTo(event2.getStart()));
 
-        LocalDateTime startTime = events.get(0).getEventStartAt();
-        LocalDateTime endTime = events.get(0).getEventEndAt();
+        LocalDateTime startTime = events.get(0).getStart();
+        LocalDateTime endTime = events.get(0).getEnd();
 
         for (int i = 1; i < events.size(); i++) {
-            LocalDateTime nextStartTime = events.get(i).getEventStartAt();
-            LocalDateTime nextEndTime = events.get(i).getEventEndAt();
+            LocalDateTime nextStartTime = events.get(i).getStart();
+            LocalDateTime nextEndTime = events.get(i).getEnd();
 
             if (endTime.isBefore(nextStartTime)) {
                 // 이벤트 간의 빈 구간이 존재
                 missingTimeIntervals.add(TeamEventsResponseDto.builder()
-                        .eventStartAt(endTime)
-                        .eventEndAt(nextStartTime)
+                        .start(endTime)
+                        .end(nextStartTime)
                         .build());
             }
 
@@ -132,7 +167,7 @@ public class TeamService {
 
     public List<TeamEventsResponseDto> deleteSleepTime(List<TeamEventsResponseDto> results, Long groupId) {
         SleepTimeDto sleepTime = getSleep(groupId);
-        LocalTime startSleep = sleepTime.getStartAt();LocalTime endSleep = sleepTime.getEndAt();
+        LocalTime startSleep = sleepTime.getStartAt(); LocalTime endSleep = sleepTime.getEndAt();
         //각 날짜별로 가장 이른 후보시간의 index 불러옴
         List<Integer> earliestTimes = getEarliestTimes(results);
         //각 날짜별로 가장 늦은 후보시간의 index 불러옴
@@ -142,19 +177,30 @@ public class TeamService {
         //그 후보시간의 startAt.isBefore sleepTime의 endAt
         for (int i = 0; i < results.size(); i++) {
             if(earliestTimes.contains(i)) {
-                if (results.get(i).getEventStartAt().toLocalTime().isBefore(endSleep)) {
+                if (results.get(i).getStart().toLocalTime().isBefore(endSleep)) {
                     //-> 후보시간.startAt = sleepTime.endAt
-                    results.get(i).setEventStartAt(LocalDateTime.of(results.get(i).getEventStartAt().toLocalDate(), endSleep));
+                    results.get(i).setStart(LocalDateTime.of(results.get(i).getStart().toLocalDate(), endSleep));
                 }
             }
         }
-
         // OR 후보시간의 endAt.isAfter sleepTime의 startAt
-        for (int i = 0; i < results.size(); i++) {
+        int size = results.size();
+        for (int i = 0; i < size; i++) {
             if(lastTimes.contains(i)) {
-                if (results.get(i).getEventEndAt().toLocalTime().isAfter(startSleep)) {
+                TeamEventsResponseDto dto = TeamEventsResponseDto.builder()
+                        .start(LocalDateTime.of(results.get(i+1).getStart().toLocalDate(),endSleep))
+                        .end(LocalDateTime.of(results.get(i+1).getStart().toLocalDate(),results.get(i).getEnd().toLocalTime()))
+                        .build();
+                results.add(dto);
+            }
+        }
+        lastTimes = getLastTimes(results);
+        size = results.size();
+        for (int i = 0; i < size; i++) {
+            if(lastTimes.contains(i)) {
+                if (results.get(i).getEnd().toLocalTime().isAfter(startSleep)) {
                     //-> 후보시간.endAt = sleepTime.startAt
-                    results.get(i).setEventEndAt(LocalDateTime.of(results.get(i).getEventEndAt().toLocalDate(), startSleep));
+                    results.get(i).setEnd(LocalDateTime.of(results.get(i).getEnd().toLocalDate(), startSleep));
                 }
             }
         }
@@ -163,49 +209,55 @@ public class TeamService {
 
     private List<Integer> getLastTimes(List<TeamEventsResponseDto> results) {
         List<Integer> lastTimes = results.stream()
-                .collect(Collectors.groupingBy(dto -> dto.getEventStartAt().toLocalDate(),
-                        Collectors.maxBy(Comparator.comparing(TeamEventsResponseDto::getEventEndAt))))
+                .collect(Collectors.groupingBy(dto -> dto.getStart().toLocalDate(),
+                        Collectors.maxBy(Comparator.comparing(TeamEventsResponseDto::getEnd))))
                 .values()
                 .stream()
                 .map(opt -> opt.orElse(null))
                 .map(results::indexOf)
                 .collect(Collectors.toList());
-        for(Integer i :lastTimes) {
-            System.out.println("l : "+i);
-        }
         return lastTimes;
     }
 
     private List<Integer> getEarliestTimes(List<TeamEventsResponseDto> results) {
         List<Integer> earliestTimes = results.stream()
-                .collect(Collectors.groupingBy(dto -> dto.getEventStartAt().toLocalDate(),
-                        Collectors.minBy(Comparator.comparing(TeamEventsResponseDto::getEventStartAt))))
+                .collect(Collectors.groupingBy(dto -> dto.getStart().toLocalDate(),
+                        Collectors.minBy(Comparator.comparing(TeamEventsResponseDto::getStart))))
                 .values()
                 .stream()
                 .map(opt -> opt.orElse(null))
                 .map(results::indexOf)
                 .collect(Collectors.toList());
-        for(Integer i : earliestTimes) {
-            System.out.println("e : "+i);
-        }
         return earliestTimes;
     }
 
     public List<TeamEventsResponseDto> setFirstLast (List < TeamEventsResponseDto > results, List < TeamEventsResponseDto > events, Long groupId){
         TeamEventsResponseDto dtoSetFirst = TeamEventsResponseDto.builder()
-                .eventStartAt(LocalDateTime.of(events.get(0).getEventStartAt().toLocalDate(), LocalTime.of(0, 0)))
-                .eventEndAt(events.get(0).getEventStartAt())
+                .start(LocalDateTime.of(events.get(0).getStart().toLocalDate(), LocalTime.of(0, 0)))
+                .end(events.get(0).getStart())
                 .build();
         results.add(0, dtoSetFirst);
 
         int lastIdx = events.size() - 1;
         TeamEventsResponseDto dtoSetLast = TeamEventsResponseDto.builder()
-                .eventStartAt(events.get(lastIdx).getEventEndAt())
-                .eventEndAt(LocalDateTime.of(events.get(lastIdx).getEventStartAt().toLocalDate().plusDays(1), LocalTime.of(0, 0)))
+                .start(events.get(lastIdx).getEnd())
+                .end(LocalDateTime.of(events.get(lastIdx).getStart().toLocalDate().plusDays(1), LocalTime.of(0, 0)))
                 .build();
         results.add(dtoSetLast);
         return results;
 
+    }
+    public List<TeamEventsResponseDto> deleteMinuteTime(List<TeamEventsResponseDto> results, int minute) {
+
+        Iterator<TeamEventsResponseDto> iterator = results.iterator();
+        while (iterator.hasNext()) {
+            TeamEventsResponseDto item = iterator.next();
+            Duration duration = Duration.between(item.getStart(), item.getEnd());
+            if (minute>duration.toMinutes()) {
+                iterator.remove(); // 요소 삭제
+            }
+        }
+        return results;
     }
 
     public Long getTeamId(String teamName){
